backend/given_images/requirements.txt
selenium
Pillow
requests
webdriver_manager
urllib3